chore: refactor to shared function transposeIntervalSetByDegree
