+++ packages/scale/index.ts
@@ -5,13 +5,14 @@
  */
 import { all as chordTypes } from "@tonaljs/chord-type";
 import { range as nums, rotate } from "@tonaljs/collection";
-import { deprecate, note, NoteName, transpose } from "@tonaljs/core";
 import {
-  enharmonic,
-  fromMidi,
-  sortedUniqNames,
-  transposeOctaves,
-} from "@tonaljs/note";
+  deprecate,
+  note,
+  NoteName,
+  transpose,
+  transposeIntervalSetByDegree,
+} from "@tonaljs/core";
+import { enharmonic, fromMidi, sortedUniqNames } from "@tonaljs/note";
 import { isSubsetOf, isSupersetOf, modes } from "@tonaljs/pcset";
 import {
   all as scaleTypes,
@@ -245,17 +246,8 @@ export function rangeOf(scale: string | string[]) {
 export function degrees(scaleName: string | ScaleNameTokens) {
   const scale = get(scaleName);
   const intervals = scale.intervals;
-  const len = intervals.length;
   const tonic = scale.tonic;
-  return (degree: number) => {
-    if (!tonic || degree === 0) return "";
-    const normalized = degree > 0 ? degree - 1 : degree;
-    const index =
-      normalized < 0 ? (len - (-normalized % len)) % len : normalized % len;
-    const octaves = Math.floor(normalized / len);
-    const root = transposeOctaves(tonic, octaves);
-    return transpose(root, intervals[index]);
-  };
+  return transposeIntervalSetByDegree(intervals, tonic ?? "");
 }
 
 export default {
